Implement commit re-ordering for fixup and squash
If fixup or squash is specified for a commit, it will be moved onto the most recently added commit that matches.
If no matching commit is found, the program exits with an error.
If a commit is already marked as fixup or squash upon arrival, it is NOT moved, but rather left in place.

TODO:
handle default squash or fixup in 2 ways:
normal mode - commits which don't fixup or squash a recognizable commit are simply picked
            - instead of resulting in an error
flatten mode - the first commit is picked, and all others are simply set to fixup/squash
             - but not moved so the entire commit history will be compacted

Also, I need to write a test or two.