rewrite parseInput().
The primary difference is in the way fixup commits that need to be moved are handled.

Summary:
If a commit is marked for fixup in the configuration, and is already a fixup in the rebase-todo,
  it is placed immediately following the commit placed just before (even if that commit ended up somewhere else).
If a commit is marked for fixup in the configuration, but NOT in the rebase-todo,
  its commit message is used to look up the commit to attach to. If none can be found, this is an error, but
  if the commit is not a fixup commit, it will be applied at HEAD.
If a commit is marked for something other than fixup in the configuration, but fixup in rebase-todo,
  it will stay in its current place and lose its fixup status.

These also apply to squash commits, which are interchangeable and interleavable with fixups.
The two are applied to other commits in the order they appear.